Basic Implementation
Not sure if we need the node-modules folder and package-lock.json file that are outside of the a2-react-app folder, those were created when i was trying to do everything manually, but i dont know if they are necessary since i implemented everything correctly when i followed the lab instructions. let me know what you think.